+++ multipicker/src/main/java/im/vector/lib/multipicker/entity/MultiPickerMediaType.kt
@@ -16,15 +16,8 @@
 
 package im.vector.lib.multipicker.entity
 
-import android.net.Uri
-
-data class MultiPickerVideoType(
-        override val displayName: String?,
-        override val size: Long,
-        override val mimeType: String?,
-        override val contentUri: Uri,
-        override val width: Int,
-        override val height: Int,
-        override val orientation: Int,
-        val duration: Long
-) : MultiPickerBaseMediaType
+interface MultiPickerBaseMediaType : MultiPickerBaseType {
+    val width: Int
+    val height: Int
+    val orientation: Int
+}