+++ CHANGES.md
@@ -9,6 +9,7 @@ Improvements 🙌:
  - Delete and react to stickers (#3250)
  - Compress video before sending (#442)
  - Improve file too big error detection (#3245)
+ - User can now select video when selecting Gallery to send attachments to a room
 
 Bugfix 🐛:
  - Message states cosmetic changes (#3007)